backend/backend/requirements.txt
Django==2
django-cors-headers==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djoser==2.0.3
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
